+++ connectome_tools/s2f_recipe/utils.py
@@ -2,7 +2,7 @@
 import logging
 from abc import ABC, abstractmethod
 
-from connectome_tools.utils import run_parallel, run_sequential
+from connectome_tools.utils import run_parallel, run_sequential, timed
 
 L = logging.getLogger(__name__)
 
@@ -31,15 +31,16 @@ class BaseExecutor(ABC):
 
     def run(self, *args, **kwargs):
         """Run the executor."""
-        L.info("Preparing strategy '%s'...", self.name)
-        tasks = self.prepare(*args, **kwargs)
-        if self.is_parallel:
-            L.info("Running strategy '%s' in parallel...", self.name)
-            results = run_parallel(tasks, self.jobs, self.base_seed)
-        else:
-            L.info("Running strategy '%s' sequentially...", self.name)
-            results = run_sequential(tasks)
-        return results
+        with timed(L, f"Executed strategy {self.name}"):
+            L.info("Preparing strategy '%s'...", self.name)
+            tasks = self.prepare(*args, **kwargs)
+            if self.is_parallel:
+                L.info("Running strategy '%s' in parallel...", self.name)
+                results = run_parallel(tasks, self.jobs, self.base_seed)
+            else:
+                L.info("Running strategy '%s' sequentially...", self.name)
+                results = run_sequential(tasks)
+            return results
 
     @property
     def name(self):